Refactor and enhance user-related functionalities
- Removed unused fields from AiRecommendationResponse and UpdateInformationDTO.
- Introduced ResetPasswordDTO for handling password reset requests.
- Updated User entity by removing phone and interests fields.
- Added new methods in GroupMemberRepository and UserRepository for better data retrieval.
- Enhanced PostRepository by removing outdated query methods.
- Implemented password reset functionality in AuthService with email notifications.
- Created EmailService for sending password reset emails.
- Improved FollowService to include more robust user suggestion logic.
- Added unit tests for FollowService to ensure correct functionality.
- Refactored GroupService to utilize active users for group seeding.
- Updated PostService to improve post handling and user interest management.
- Removed deprecated weekly report retrieval method from ReportService.
- Enhanced UserService to manage user group memberships effectively.

+++ src/main/java/com/example/campushub/responses/admin/AdminTopicDistributionResponse.java
@@ -13,5 +13,5 @@ import lombok.Setter;
 @Builder
 public class AdminTopicDistributionResponse {
     private String label;
-    private long value;
+    private double value;
 }
